Always use the current logger in EnsureNotUsedBy
By assigning the logger to an ivar here, it gets evaluated once and
stays around for the entire lifetime of the application. If the logdev
gets closed (as it is in the test suite), it will generate a
warning. This change ensures that messages will always get written to
the current log.